Fix 64-bit index stride bug: long tagoffset should be int
The endtags struct used long tagoffset, but ReadKey/WriteKey read
and write sizeof(int) bytes. On 64-bit platforms where sizeof(long)=8,
the struct stride was 4 bytes too large, causing init_preind() to
silently skip ~25% of index entries when iterating the binary endtag
files.

+++ src/includes/endtags.h
@@ -7,7 +7,7 @@
 
 typedef struct {
 	char tagstring[MAXTAG];
-	long tagoffset;
+	int tagoffset;
 } endtags;
 
 #define tagstring_of(X) (X)->tagstring
